ext/ckanext-digitraffic_opentelemetry/requirements.txt
#typing-extensions==4.13.1
#aws-opentelemetry-distro==0.9.0

opentelemetry-sdk==1.34.0
opentelemetry-sdk-extension-aws==2.1.0
opentelemetry-propagator-aws-xray==1.0.2
opentelemetry-exporter-otlp==1.34.0
opentelemetry-instrumentation-flask==0.55b0
opentelemetry-instrumentation-requests==0.55b0
opentelemetry-instrumentation-urllib3==0.55b0
opentelemetry-instrumentation-logging==0.55b0
opentelemetry-processor-baggage==0.55b0
opentelemetry-instrumentation-sqlalchemy==0.55b0
opentelemetry-instrumentation-redis==0.55b0
opentelemetry-instrumentation-psycopg2==0.55b0
opentelemetry-instrumentation-threading==0.55b0